pyproject.toml
[tool.poetry]
name = "conference room"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-jazzmin = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
